reduce duplication, fix a bug
