+++ src/index.js
@@ -25,6 +25,8 @@ module.exports = {
 	SchemaPiece: require('./lib/settings/schema/SchemaPiece'),
 
 	// lib/structures/base
+	AliasPiece: require('./lib/structures/base/AliasPiece'),
+	AliasStore: require('./lib/structures/base/AliasStore'),
 	Piece: require('./lib/structures/base/Piece'),
 	Store: require('./lib/structures/base/Store'),
