+++ src/lib/structures/Command.js
@@ -1,5 +1,5 @@
 const { Permissions } = require('discord.js');
-const Piece = require('./base/Piece');
+const AliasPiece = require('./base/AliasPiece');
 const Usage = require('../usage/Usage');
 const CommandUsage = require('../usage/CommandUsage');
 const RateLimitManager = require('../util/RateLimitManager');
@@ -9,13 +9,12 @@ const { isFunction } = require('../util/util');
  * Base class for all Klasa Commands. See {@tutorial CreatingCommands} for more information how to use this class
  * to build custom commands.
  * @tutorial CreatingCommands
- * @extends Piece
+ * @extends AliasPiece
  */
-class Command extends Piece {
+class Command extends AliasPiece {
 
 	/**
-	 * @typedef {PieceOptions} CommandOptions
-	 * @property {string[]} [aliases=[]] Any command aliases
+	 * @typedef {AliasPieceOptions} CommandOptions
 	 * @property {boolean} [autoAliases=true] If automatic aliases should be added (adds aliases of name and aliases without dashes)
 	 * @property {external:PermissionResolvable} [requiredPermissions=0] The required Discord permissions for the bot to use this command
 	 * @property {number} [bucket=1] The number of times this command can be run before ratelimited by the cooldown
@@ -50,12 +49,6 @@ class Command extends Piece {
 
 		this.name = this.name.toLowerCase();
 
-		/**
-		 * The aliases for this command
-		 * @since 0.0.1
-		 * @type {string[]}
-		 */
-		this.aliases = options.aliases;
 		if (options.autoAliases) {
 			if (this.name.includes('-')) this.aliases.push(this.name.replace(/-/g, ''));
 			for (const alias of this.aliases) if (alias.includes('-')) this.aliases.push(alias.replace(/-/g, ''));
@@ -315,7 +308,6 @@ class Command extends Piece {
 	toJSON() {
 		return {
 			...super.toJSON(),
-			aliases: this.aliases.slice(0),
 			requiredPermissions: this.requiredPermissions.toArray(false),
 			bucket: this.bucket,
 			category: this.category,
